some bugs fixed

+++ ArticleManagement/ArticleManagement/src/main/java/nwtprojekat/ArticleManagement/service/ArticleService.java
@@ -1,5 +1,6 @@
 package nwtprojekat.ArticleManagement.service;
 
+import nwtprojekat.ArticleManagement.exception.ArticleNotFoundException;
 import nwtprojekat.ArticleManagement.model.Article;
 import nwtprojekat.ArticleManagement.repository.ArticleRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,20 +14,36 @@ public class ArticleService {
     @Autowired
     private ArticleRepository articleRepository;
 
-    public Optional<Article> getArticleById(Long id) {
-        return articleRepository.findById(id);
+    public Article getArticleById(Long id) {
+        return articleRepository.findById(id)
+                .orElseThrow(() -> new ArticleNotFoundException(id));
     }
 
     public Article createArticle(Article article) {
         return articleRepository.save(article);
     }
 
-    public void deleteArticle(Long id) {
+    public void deleteArticleById(Long id) {
+        if (!articleRepository.existsById(id)) {
+            throw new ArticleNotFoundException(id);
+        }
         articleRepository.deleteById(id);
+    }
+
+    public boolean existsArticleById(Long id) {
+        return articleRepository.existsById(id);
+    }
+
+    public Article updateArticle(Article updatedArticle) {
+        return articleRepository.save(updatedArticle);
     }
 
     public List<Article> getAllArticles() {
         return articleRepository.findAll();
+    }
+
+    public List<Article> findArticlesByAuthor(String authorName) {
+        return articleRepository.findByAuthor(authorName);
     }
 
 }
